🚧 move MainMenu to menu component folder
